Separate methods on challengeService into challengeRegistrantService and challengeSubmissionService
