+++ src/main/java/com/techlooper/controller/ChallengeSubmissionController.java
@@ -4,6 +4,7 @@ import com.techlooper.entity.ChallengeRegistrantEntity;
 import com.techlooper.entity.ChallengeSubmissionEntity;
 import com.techlooper.model.ChallengePhaseEnum;
 import com.techlooper.model.ChallengeSubmissionDto;
+import com.techlooper.service.ChallengeRegistrantService;
 import com.techlooper.service.ChallengeService;
 import com.techlooper.service.ChallengeSubmissionService;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -25,6 +26,9 @@ public class ChallengeSubmissionController {
     @Resource
     private ChallengeService challengeService;
 
+    @Resource
+    private ChallengeRegistrantService challengeRegistrantService;
+
     @RequestMapping(value = "user/challengeSubmission", method = RequestMethod.POST)
     public ChallengeSubmissionEntity submitMyResult(@RequestBody ChallengeSubmissionDto challengeSubmissionDto) {
         return challengeSubmissionService.submitMyResult(challengeSubmissionDto);
@@ -33,7 +37,7 @@ public class ChallengeSubmissionController {
     @RequestMapping(value = "user/challengeSubmissionPhase/{registrantEmail}/{challengeId}", method = RequestMethod.GET)
     public ChallengePhaseEnum challengeSubmissionPhase(@PathVariable String registrantEmail, @PathVariable Long challengeId) {
         ChallengeRegistrantEntity registrantEntity =
-                challengeService.findRegistrantByChallengeIdAndEmail(challengeId, registrantEmail);
+                challengeRegistrantService.findRegistrantByChallengeIdAndEmail(challengeId, registrantEmail);
         return (registrantEntity != null && registrantEntity.getActivePhase() != null) ?
                 registrantEntity.getActivePhase() : ChallengePhaseEnum.REGISTRATION;
     }
