+++ src/main/java/com/techlooper/service/ChallengeSubmissionService.java
@@ -4,7 +4,9 @@ import com.techlooper.entity.ChallengeSubmissionEntity;
 import com.techlooper.model.ChallengePhaseEnum;
 import com.techlooper.model.ChallengeSubmissionDto;
 import com.techlooper.model.ChallengeSubmissionPhaseItem;
+import com.techlooper.model.TimePeriodEnum;
 
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -19,4 +21,6 @@ public interface ChallengeSubmissionService {
     void markChallengeSubmissionAsRead(ChallengeSubmissionDto challengeSubmissionDto);
 
     ChallengeSubmissionEntity findChallengeSubmissionByRegistrantPhase(Long registrantId, ChallengePhaseEnum phase);
+
+    List<ChallengeSubmissionEntity> findChallengeSubmissionWithinPeriod(Long challengeId, Long currentDateTime, TimePeriodEnum period);
 }
