+++ src/main/java/com/techlooper/service/impl/ChallengeServiceImpl.java
@@ -1,13 +1,10 @@
 package com.techlooper.service.impl;
 
-import com.techlooper.dto.ChallengeQualificationDto;
 import com.techlooper.dto.EmailSettingDto;
-import com.techlooper.dto.RejectRegistrantDto;
 import com.techlooper.entity.*;
 import com.techlooper.model.*;
 import com.techlooper.repository.elasticsearch.ChallengeRegistrantRepository;
 import com.techlooper.repository.elasticsearch.ChallengeRepository;
-import com.techlooper.repository.elasticsearch.ChallengeSubmissionRepository;
 import com.techlooper.service.*;
 import com.techlooper.util.DataUtils;
 import freemarker.template.Template;
@@ -20,16 +17,10 @@ import org.elasticsearch.index.query.*;
 import org.elasticsearch.search.aggregations.Aggregations;
 import org.elasticsearch.search.aggregations.metrics.sum.Sum;
 import org.elasticsearch.search.aggregations.metrics.sum.SumBuilder;
-import org.elasticsearch.search.sort.SortOrder;
-import org.joda.time.DateTime;
-import org.joda.time.format.DateTimeFormat;
-import org.joda.time.format.DateTimeFormatter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.elasticsearch.core.ElasticsearchTemplate;
-import org.springframework.data.elasticsearch.core.FacetedPage;
 import org.springframework.data.elasticsearch.core.query.NativeSearchQueryBuilder;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
@@ -47,15 +38,15 @@ import java.io.StringWriter;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.*;
-import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
+import static com.techlooper.model.ChallengePhaseEnum.*;
 import static com.techlooper.util.DateTimeUtils.*;
-import static java.util.stream.Collectors.*;
+import static java.util.stream.Collectors.joining;
+import static java.util.stream.Collectors.toList;
 import static org.elasticsearch.index.query.FilterBuilders.*;
 import static org.elasticsearch.index.query.QueryBuilders.*;
 import static org.elasticsearch.search.aggregations.AggregationBuilders.sum;
-import static org.elasticsearch.search.sort.SortBuilders.fieldSort;
 
 /**
  * Created by NguyenDangKhoa on 6/29/15.
@@ -65,13 +56,7 @@ public class ChallengeServiceImpl implements ChallengeService {
 
     private final static Logger LOGGER = LoggerFactory.getLogger(ChallengeServiceImpl.class);
 
-    private final static ChallengePhaseEnum CHALLENGE_TIMELINE[] = {
-            ChallengePhaseEnum.FINAL,
-            ChallengePhaseEnum.PROTOTYPE,
-            ChallengePhaseEnum.UIUX,
-            ChallengePhaseEnum.IDEA,
-            ChallengePhaseEnum.REGISTRATION
-    };
+    private final static ChallengePhaseEnum CHALLENGE_TIMELINE[] = {FINAL, PROTOTYPE, UIUX, IDEA, REGISTRATION};
 
     @Resource
     private ElasticsearchTemplate elasticsearchTemplateUserImport;
@@ -118,12 +103,6 @@ public class ChallengeServiceImpl implements ChallengeService {
     @Value("${mail.confirmUserJoinChallenge.subject.en}")
     private String confirmUserJoinChallengeMailSubjectEn;
 
-    @Resource
-    private Template alertEmployerChallengeMailTemplateEn;
-
-    @Resource
-    private Template alertEmployerChallengeMailTemplateVi;
-
     @Value("${mail.alertEmployerChallenge.subject.vn}")
     private String alertEmployerChallengeMailSubjectVn;
 
@@ -141,9 +120,6 @@ public class ChallengeServiceImpl implements ChallengeService {
 
     @Resource
     private ChallengeRegistrantRepository challengeRegistrantRepository;
-
-    @Resource
-    private ChallengeSubmissionRepository challengeSubmissionRepository;
 
     @Resource
     private Mapper dozerMapper;
@@ -270,19 +246,19 @@ public class ChallengeServiceImpl implements ChallengeService {
         templateModel.put("challengeNameAlias", challengeEntity.getChallengeName().replaceAll("\\W", "-"));
 
         int numberOfDays = 0;
-        if (challengePhase == ChallengePhaseEnum.REGISTRATION) {
+        if (challengePhase == REGISTRATION) {
             numberOfDays = daysBetween(currentDate(), challengeEntity.getRegistrationDateTime()) + 1;
-        } else if (challengePhase == ChallengePhaseEnum.IN_PROGRESS) {
+        } else if (challengePhase == IN_PROGRESS) {
             numberOfDays = daysBetween(currentDate(), challengeEntity.getSubmissionDateTime()) + 1;
-        } else if (challengePhase == ChallengePhaseEnum.IDEA) {
+        } else if (challengePhase == IDEA) {
             numberOfDays = daysBetween(currentDate(), challengeEntity.getIdeaSubmissionDateTime()) + 1;
-        } else if (challengePhase == ChallengePhaseEnum.UIUX) {
+        } else if (challengePhase == UIUX) {
             numberOfDays = daysBetween(currentDate(), challengeEntity.getUxSubmissionDateTime()) + 1;
-        } else if (challengePhase == ChallengePhaseEnum.PROTOTYPE) {
+        } else if (challengePhase == PROTOTYPE) {
             numberOfDays = daysBetween(currentDate(), challengeEntity.getPrototypeSubmissionDateTime()) + 1;
-        } else if (challengePhase == ChallengePhaseEnum.IN_PROGRESS) {
+        } else if (challengePhase == IN_PROGRESS) {
             numberOfDays = daysBetween(currentDate(), challengeEntity.getSubmissionDateTime()) + 1;
-        } else if (challengePhase == ChallengePhaseEnum.FINAL) {
+        } else if (challengePhase == FINAL) {
             numberOfDays = daysBetween(currentDate(), challengeEntity.getSubmissionDateTime()) + 1;
         }
 
@@ -351,24 +327,18 @@ public class ChallengeServiceImpl implements ChallengeService {
     private String getNotifyRegistrantChallengeTimelineSubject(
             ChallengeRegistrantEntity challengeRegistrantEntity, ChallengePhaseEnum challengePhase) {
         if (challengeRegistrantEntity.getLang() == Language.vi) {
-            if (challengePhase == ChallengePhaseEnum.REGISTRATION) {
+            if (challengePhase == REGISTRATION) {
                 return notifyChallengeTimelineRegistrationMailSubjectVn;
             } else {
                 return notifyChallengeTimelineInProgressMailSubjectVn;
             }
         } else {
-            if (challengePhase == ChallengePhaseEnum.REGISTRATION) {
+            if (challengePhase == REGISTRATION) {
                 return notifyChallengeTimelineRegistrationMailSubjectEn;
             } else {
                 return notifyChallengeTimelineInProgressMailSubjectEn;
             }
         }
-    }
-
-    public Long getNumberOfRegistrants(Long challengeId) {
-        NativeSearchQueryBuilder searchQueryBuilder = new NativeSearchQueryBuilder().withSearchType(SearchType.COUNT);
-        searchQueryBuilder.withFilter(FilterBuilders.termFilter("challengeId", challengeId));
-        return challengeRegistrantRepository.search(searchQueryBuilder.build()).getTotalElements();
     }
 
     public void sendApplicationEmailToContestant(ChallengeEntity challengeEntity, ChallengeRegistrantEntity challengeRegistrantEntity) throws MessagingException, IOException, TemplateException {
@@ -381,20 +351,10 @@ public class ChallengeServiceImpl implements ChallengeService {
         sendContestApplicationEmail(template, mailSubject, emailAddress, challengeEntity, challengeRegistrantEntity, false);
     }
 
-    public void sendApplicationEmailToEmployer(ChallengeEntity challengeEntity, ChallengeRegistrantEntity challengeRegistrantEntity) throws MessagingException, IOException, TemplateException {
-        Template template = challengeRegistrantEntity.getLang() == Language.vi ?
-                alertEmployerChallengeMailTemplateVi : alertEmployerChallengeMailTemplateEn;
-        String mailSubject = challengeRegistrantEntity.getLang() == Language.vi ?
-                alertEmployerChallengeMailSubjectVn : alertEmployerChallengeMailSubjectEn;
-        mailSubject = String.format(mailSubject, challengeEntity.getChallengeName());
-        Address[] emailAddress = getRecipientAddresses(challengeEntity, false);
-        sendContestApplicationEmail(template, mailSubject, emailAddress, challengeEntity, challengeRegistrantEntity, true);
-    }
-
     public long joinChallenge(ChallengeRegistrantDto challengeRegistrantDto) {
         ChallengeRegistrantEntity entity = joinChallengeEntity(challengeRegistrantDto);
         if (entity != null) {
-            return getNumberOfRegistrants(challengeRegistrantDto.getChallengeId());
+            return challengeRegistrantService.getNumberOfRegistrants(challengeRegistrantDto.getChallengeId());
         }
         return 0;
     }
@@ -433,7 +393,8 @@ public class ChallengeServiceImpl implements ChallengeService {
         ArrayList<ChallengeDetailDto> challenges = new ArrayList<>();
         challengeIterator.forEach(challengeEntity -> {
             ChallengeDetailDto challengeDetailDto = dozerMapper.map(challengeEntity, ChallengeDetailDto.class);
-            challengeDetailDto.setNumberOfRegistrants(getNumberOfRegistrants(challengeEntity.getChallengeId()));
+            challengeDetailDto.setNumberOfRegistrants(challengeRegistrantService.getNumberOfRegistrants(
+                    challengeEntity.getChallengeId()));
             challenges.add(challengeDetailDto);
         });
         return sortChallengesByDescendingStartDate(challenges);
@@ -555,22 +516,8 @@ public class ChallengeServiceImpl implements ChallengeService {
         }).collect(toList());
     }
 
-    @Override
-    public ChallengeRegistrantEntity findRegistrantByChallengeIdAndEmail(Long challengeId, String email) {
-        NativeSearchQueryBuilder searchQueryBuilder = new NativeSearchQueryBuilder().withTypes("challengeRegistrant");
-        searchQueryBuilder.withQuery(boolQuery()
-                .must(termQuery("registrantEmail", email))
-                .must(termQuery("challengeId", challengeId)));
-
-        List<ChallengeRegistrantEntity> registrantEntities = DataUtils.getAllEntities(challengeRegistrantRepository, searchQueryBuilder);
-        if (!registrantEntities.isEmpty()) {
-            return registrantEntities.get(0);
-        }
-        return null;
-    }
-
     public boolean checkIfChallengeRegistrantExist(Long challengeId, String email) {
-        return findRegistrantByChallengeIdAndEmail(challengeId, email) != null;
+        return challengeRegistrantService.findRegistrantByChallengeIdAndEmail(challengeId, email) != null;
     }
 
     @Override
@@ -592,49 +539,8 @@ public class ChallengeServiceImpl implements ChallengeService {
     }
 
     @Override
-    public Long getTotalNumberOfRegistrants() {
-        return challengeRegistrantRepository.count();
-    }
-
-    @Override
     public ChallengeDetailDto getTheLatestChallenge() {
-//    NativeSearchQueryBuilder searchQueryBuilder = new NativeSearchQueryBuilder();
-//    searchQueryBuilder.withQuery(QueryBuilders.matchAllQuery());
-//    searchQueryBuilder.withSort(SortBuilders.fieldSort("challengeId").order(SortOrder.DESC));
-//    searchQueryBuilder.withPageable(new PageRequest(0, 1));
-//
-//    List<ChallengeEntity> challengeEntities = challengeRepository.search(searchQueryBuilder.build()).getContent();
-//    if (!challengeEntities.isEmpty()) {
-//      ChallengeEntity challengeEntity = challengeEntities.get(0);
-//      return dozerMapper.map(challengeEntity, ChallengeDetailDto.class);
-//    }
         return listChallenges().get(0);
-    }
-
-    public Collection<ChallengeDetailDto> findByOwnerAndCondition(String owner,
-                                                                  Predicate<? super ChallengeEntity> condition) {
-        NativeSearchQueryBuilder queryBuilder = new NativeSearchQueryBuilder().withIndices(techlooperIndex).withTypes("challenge");
-        QueryStringQueryBuilder query = queryStringQuery(owner).defaultField("authorEmail");
-        queryBuilder.withFilter(FilterBuilders.queryFilter(query));
-
-        int pageIndex = 0;
-        Set<ChallengeDetailDto> challenges = new HashSet<>();
-        while (true) {
-            queryBuilder.withPageable(new PageRequest(pageIndex++, 100));
-            FacetedPage<ChallengeEntity> page = challengeRepository.search(queryBuilder.build());
-            if (!page.hasContent()) {
-                break;
-            }
-
-            page.spliterator().forEachRemaining(challenge -> {
-                if (condition.test(challenge)) {
-                    ChallengeDetailDto challengeDetailDto = dozerMapper.map(challenge, ChallengeDetailDto.class);
-                    challengeDetailDto.setNumberOfRegistrants(countRegistrantsByChallengeId(challenge.getChallengeId()));
-                    challenges.add(challengeDetailDto);
-                }
-            });
-        }
-        return challenges;
     }
 
     public List<ChallengeDetailDto> listChallenges(String ownerEmail) {
@@ -648,29 +554,11 @@ public class ChallengeServiceImpl implements ChallengeService {
 
         for (ChallengeEntity challengeEntity : challengeEntities) {
             ChallengeDetailDto challengeDetailDto = dozerMapper.map(challengeEntity, ChallengeDetailDto.class);
-            challengeDetailDto.setNumberOfRegistrants(getNumberOfRegistrants(challengeDetailDto.getChallengeId()));
+            challengeDetailDto.setNumberOfRegistrants(challengeRegistrantService.getNumberOfRegistrants(
+                    challengeDetailDto.getChallengeId()));
             result.add(challengeDetailDto);
         }
         return result;
-    }
-
-    public Collection<ChallengeDetailDto> findInProgressChallenges(String owner) {
-        DateTimeFormatter dateTimeFormatter = DateTimeFormat.forPattern("dd/MM/yyyy");
-        return findByOwnerAndCondition(owner, challengeEntity -> {
-            DateTime startDate = dateTimeFormatter.parseDateTime(challengeEntity.getStartDateTime());
-            DateTime submissionDate = dateTimeFormatter.parseDateTime(challengeEntity.getSubmissionDateTime());
-            DateTime now = DateTime.now();
-            boolean inRange = now.isAfter(startDate) && now.isBefore(submissionDate);
-            boolean atBoundary = now.isEqual(startDate) || now.isEqual(submissionDate);
-            return inRange || atBoundary;
-        });
-    }
-
-    public Long countRegistrantsByChallengeId(Long challengeId) {
-        NativeSearchQueryBuilder queryBuilder = new NativeSearchQueryBuilder().withIndices(techlooperIndex).withTypes("challengeRegistrant");
-        queryBuilder.withFilter(FilterBuilders.queryFilter(termQuery("challengeId", challengeId)))
-                .withSearchType(SearchType.COUNT);
-        return challengeRegistrantRepository.search(queryBuilder.build()).getTotalElements();
     }
 
     public boolean delete(Long id, String ownerEmail) {
@@ -689,129 +577,6 @@ public class ChallengeServiceImpl implements ChallengeService {
             challenge.setCriteria(null);
         }
         return dozerMapper.map(challenge, ChallengeDto.class);
-    }
-
-    public Set<ChallengeRegistrantDto> findRegistrantsByOwner(RegistrantFilterCondition condition) throws ParseException {
-        Set<ChallengeRegistrantDto> result = new HashSet<>();
-        Long challengeId = condition.getChallengeId();
-        ChallengeEntity challengeEntity = challengeRepository.findOne(challengeId);
-
-        if (challengeEntity != null && challengeEntity.getAuthorEmail().equals(condition.getAuthorEmail())) {
-            List<ChallengeRegistrantEntity> registrants = filterChallengeRegistrantByDate(condition);
-            for (ChallengeRegistrantEntity registrant : registrants) {
-                ChallengeRegistrantDto registrantDto = dozerMapper.map(registrant, ChallengeRegistrantDto.class);
-                registrantDto.setSubmissions(findChallengeSubmissionByRegistrant(challengeId, registrant.getRegistrantId()));
-                result.add(registrantDto);
-            }
-        }
-
-        return result;
-    }
-
-    public ChallengeRegistrantDto saveRegistrant(String ownerEmail, ChallengeRegistrantDto challengeRegistrantDto) {
-        ChallengeRegistrantDto resultChallengeRegistrantDto = challengeRegistrantDto;
-        ChallengeEntity challenge = challengeRepository.findOne(challengeRegistrantDto.getChallengeId());
-        if (ownerEmail.equalsIgnoreCase(challenge.getAuthorEmail())) {
-            ChallengeRegistrantEntity registrant = challengeRegistrantRepository.findOne(challengeRegistrantDto.getRegistrantId());
-            challengeRegistrantDto.setRegistrantEmail(registrant.getRegistrantEmail());
-            dozerMapper.map(challengeRegistrantDto, registrant);
-            registrant = challengeRegistrantRepository.save(registrant);
-            resultChallengeRegistrantDto = dozerMapper.map(registrant, ChallengeRegistrantDto.class);
-        }
-        return resultChallengeRegistrantDto;
-    }
-
-    @Override
-    public List<ChallengeRegistrantEntity> findChallengeRegistrantWithinPeriod(
-            Long challengeId, Long currentDateTime, TimePeriodEnum period) {
-        NativeSearchQueryBuilder searchQueryBuilder = new NativeSearchQueryBuilder().withTypes("challengeRegistrant");
-
-        BoolQueryBuilder boolQueryBuilder = boolQuery();
-        boolQueryBuilder.must(termQuery("challengeId", challengeId));
-
-        Long pastTime = currentDateTime - period.getMiliseconds() > 0 ? currentDateTime - period.getMiliseconds() : 0;
-        boolQueryBuilder.must(rangeQuery("registrantId").from(pastTime));
-        searchQueryBuilder.withQuery(boolQueryBuilder);
-        searchQueryBuilder.withSort(fieldSort("registrantId").order(SortOrder.DESC));
-        return DataUtils.getAllEntities(challengeRegistrantRepository, searchQueryBuilder);
-    }
-
-    @Override
-    public List<ChallengeRegistrantEntity> filterChallengeRegistrantByDate(RegistrantFilterCondition condition) throws ParseException {
-        List<ChallengeRegistrantEntity> result = new ArrayList<>();
-
-        if (ChallengeRegistrantFilterTypeEnum.BY_SUBMISSION.getValue().equals(condition.getFilterType())) {
-            Set<Long> registrantIds = findRegistrantByChallengeSubmissionDate(
-                    condition.getChallengeId(), condition.getFromDate(), condition.getToDate());
-            for (Long registrantId : registrantIds) {
-                ChallengeRegistrantEntity registrantEntity = challengeRegistrantRepository.findOne(registrantId);
-                if (registrantEntity != null) {
-                    result.add(registrantEntity);
-                }
-            }
-
-            if (StringUtils.isNotEmpty(condition.getPhase()) &&
-                    !ChallengePhaseEnum.ALL_PHASES.getValue().equals(condition.getPhase())) {
-                result = result.stream().filter(registrantEntity ->
-                        condition.getPhase().equals(registrantEntity.getActivePhase().getValue())).collect(toList());
-            }
-        } else {
-            NativeSearchQueryBuilder searchQueryBuilder = new NativeSearchQueryBuilder().withTypes("challengeRegistrant");
-            BoolQueryBuilder boolQueryBuilder = QueryBuilders.boolQuery();
-            final String registrationPhase = ChallengePhaseEnum.REGISTRATION.getValue();
-
-            if (condition.getChallengeId() != null) {
-                boolQueryBuilder.must(termQuery("challengeId", condition.getChallengeId()));
-            }
-
-            if (StringUtils.isNotEmpty(condition.getPhase())
-                    && !registrationPhase.equals(condition.getPhase())
-                    && !ChallengePhaseEnum.ALL_PHASES.getValue().equals(condition.getPhase())) {
-                boolQueryBuilder.must(matchPhraseQuery("activePhase", condition.getPhase()));
-            }
-
-            if (StringUtils.isNotEmpty(condition.getFilterType()) &&
-                    (StringUtils.isNotEmpty(condition.getFromDate()) || StringUtils.isNotEmpty(condition.getToDate()))) {
-                RangeQueryBuilder rangeQueryBuilder = QueryBuilders.rangeQuery(condition.getFilterType());
-
-                if (StringUtils.isNotEmpty(condition.getFromDate())) {
-                    Long from = string2Date(condition.getFromDate(), BASIC_DATE_PATTERN).getTime();
-                    rangeQueryBuilder.from(from);
-                }
-                if (StringUtils.isNotEmpty(condition.getToDate())) {
-                    Long to = string2Date(condition.getToDate(), BASIC_DATE_PATTERN).getTime() +
-                            TimePeriodEnum.TWENTY_FOUR_HOURS.getMiliseconds();
-                    rangeQueryBuilder.to(to);
-                }
-
-                boolQueryBuilder.must(rangeQueryBuilder);
-            }
-
-            searchQueryBuilder.withQuery(boolQueryBuilder);
-            result.addAll(DataUtils.getAllEntities(challengeRegistrantRepository, searchQueryBuilder));
-
-            if (registrationPhase.equals(condition.getPhase())) {
-                result = result.stream().filter(registrantEntity -> registrantEntity.getActivePhase() == null
-                        || registrationPhase.equals(registrantEntity.getActivePhase().getValue())).collect(toList());
-            }
-        }
-
-        return result;
-    }
-
-    @Override
-    public List<ChallengeSubmissionEntity> findChallengeSubmissionWithinPeriod(
-            Long challengeId, Long currentDateTime, TimePeriodEnum period) {
-        NativeSearchQueryBuilder searchQueryBuilder = new NativeSearchQueryBuilder().withTypes("challengeSubmission");
-
-        BoolQueryBuilder boolQueryBuilder = boolQuery();
-        boolQueryBuilder.must(termQuery("challengeId", challengeId));
-
-        boolQueryBuilder.must(rangeQuery("submissionDateTime").from(yesterdayDate()));
-        searchQueryBuilder.withQuery(boolQueryBuilder);
-        searchQueryBuilder.withSort(fieldSort("challengeSubmissionId").order(SortOrder.DESC));
-
-        return DataUtils.getAllEntities(challengeSubmissionRepository, searchQueryBuilder);
     }
 
     @Override
@@ -853,12 +618,12 @@ public class ChallengeServiceImpl implements ChallengeService {
         templateModel.put("yesterdayDateTime", yesterdayDate());
 
         Long currentDateTime = new Date().getTime();
-        List<ChallengeRegistrantEntity> latestRegistrants = findChallengeRegistrantWithinPeriod(
+        List<ChallengeRegistrantEntity> latestRegistrants = challengeRegistrantService.findChallengeRegistrantWithinPeriod(
                 challengeEntity.getChallengeId(), currentDateTime, TimePeriodEnum.TWENTY_FOUR_HOURS);
         templateModel.put("numberOfRegistrants", latestRegistrants.size());
         templateModel.put("latestRegistrants", latestRegistrants);
 
-        List<ChallengeSubmissionEntity> latestSubmissions = findChallengeSubmissionWithinPeriod(
+        List<ChallengeSubmissionEntity> latestSubmissions = challengeSubmissionService.findChallengeSubmissionWithinPeriod(
                 challengeEntity.getChallengeId(), currentDateTime, TimePeriodEnum.TWENTY_FOUR_HOURS);
         templateModel.put("numberOfSubmissions", latestSubmissions.size());
         templateModel.put("latestSubmissions", latestSubmissions);
@@ -888,7 +653,7 @@ public class ChallengeServiceImpl implements ChallengeService {
 
     public boolean sendEmailToDailyChallengeRegistrants(String challengeOwner, Long challengeId, Long now, EmailContent emailContent) {
         if (isOwnerOfChallenge(challengeOwner, challengeId)) {
-            List<ChallengeRegistrantEntity> registrants = findChallengeRegistrantWithinPeriod(challengeId, now, TimePeriodEnum.TWENTY_FOUR_HOURS);
+            List<ChallengeRegistrantEntity> registrants = challengeRegistrantService.findChallengeRegistrantWithinPeriod(challengeId, now, TimePeriodEnum.TWENTY_FOUR_HOURS);
             String csvEmails = registrants.stream().map(ChallengeRegistrantEntity::getRegistrantEmail).distinct().collect(joining(","));
             try {
                 emailContent.setRecipients(InternetAddress.parse(csvEmails));
@@ -1014,18 +779,6 @@ public class ChallengeServiceImpl implements ChallengeService {
     }
 
     @Override
-    public List<ChallengeSubmissionDto> findChallengeSubmissionByRegistrant(Long challengeId, Long registrantId) {
-        NativeSearchQueryBuilder searchQueryBuilder = new NativeSearchQueryBuilder().withTypes("challengeSubmission");
-        BoolFilterBuilder boolFilterBuilder = new BoolFilterBuilder();
-        boolFilterBuilder.must(termFilter("challengeId", challengeId));
-        boolFilterBuilder.must(termFilter("registrantId", registrantId));
-
-        searchQueryBuilder.withQuery(filteredQuery(matchAllQuery(), boolFilterBuilder));
-        List<ChallengeSubmissionEntity> submissions = DataUtils.getAllEntities(challengeSubmissionRepository, searchQueryBuilder);
-        return submissions.stream().map(submission -> dozerMapper.map(submission, ChallengeSubmissionDto.class)).collect(toList());
-    }
-
-    @Override
     public void updateSendEmailToContestantResultCode(ChallengeRegistrantEntity challengeRegistrantEntity, EmailSentResultEnum code) {
         if (challengeRegistrantEntity != null) {
             challengeRegistrantEntity.setLastEmailSentDateTime(currentDate(BASIC_DATE_TIME_PATTERN));
@@ -1042,33 +795,9 @@ public class ChallengeServiceImpl implements ChallengeService {
         }
     }
 
-    public Set<Long> findRegistrantByChallengeSubmissionDate(Long challengeId, String fromDate, String toDate) {
-        NativeSearchQueryBuilder searchQueryBuilder = new NativeSearchQueryBuilder().withTypes("challengeSubmission");
-        BoolQueryBuilder boolQueryBuilder = boolQuery();
-        boolQueryBuilder.must(termQuery("challengeId", challengeId));
-
-        RangeQueryBuilder submissionDateQuery = QueryBuilders.rangeQuery("submissionDateTime");
-        if (StringUtils.isNotEmpty(fromDate) || StringUtils.isNotEmpty(toDate)) {
-            if (StringUtils.isNotEmpty(fromDate)) {
-                submissionDateQuery.from(fromDate);
-            }
-
-            if (StringUtils.isNotEmpty(toDate)) {
-                submissionDateQuery.to(toDate);
-            }
-
-            boolQueryBuilder.must(submissionDateQuery);
-        }
-
-        searchQueryBuilder.withQuery(boolQueryBuilder);
-        List<ChallengeSubmissionEntity> submissions = DataUtils.getAllEntities(challengeSubmissionRepository, searchQueryBuilder);
-        return submissions.stream().map(submission -> submission.getRegistrantId()).collect(toSet());
-    }
-
     public ChallengeDetailDto getChallengeDetail(Long challengeId, String loginEmail) {
         NativeSearchQueryBuilder searchQueryBuilder = new NativeSearchQueryBuilder().withTypes("challenge");
         TermQueryBuilder challengeIdQuery = termQuery("challengeId", challengeId);
-        //MatchQueryBuilder authorEmailQuery = matchQuery("authorEmail", loginEmail).minimumShouldMatch("100%");
         TermQueryBuilder expiredChallengeQuery = termQuery("expired", Boolean.TRUE);
 
         searchQueryBuilder.withQuery(boolQuery().must(challengeIdQuery).mustNot(expiredChallengeQuery));
@@ -1077,14 +806,14 @@ public class ChallengeServiceImpl implements ChallengeService {
         if (!challengeEntities.isEmpty()) {
             ChallengeEntity challengeEntity = challengeEntities.get(0);
             ChallengeDetailDto challengeDetailDto = dozerMapper.map(challengeEntity, ChallengeDetailDto.class);
-            challengeDetailDto.setNumberOfRegistrants(getNumberOfRegistrants(challengeId));
+            challengeDetailDto.setNumberOfRegistrants(challengeRegistrantService.getNumberOfRegistrants(challengeId));
             calculateChallengePhases(challengeDetailDto);
             boolean isAuthor = challengeEntity.getAuthorEmail().equals(loginEmail);
             challengeDetailDto.setIsAuthor(isAuthor);
             if (!isAuthor) {
                 challengeDetailDto.setCriteria(null);
             }
-            challengeDetailDto.setPhaseItems(this.getChallengeRegistrantFunnel(challengeId, loginEmail));
+            challengeDetailDto.setPhaseItems(challengeRegistrantService.getChallengeRegistrantFunnel(challengeId, loginEmail));
 
             try {
                 Boolean isClosed = daysBetween(challengeDetailDto.getSubmissionDateTime(), currentDate()) > 0;
@@ -1124,108 +853,12 @@ public class ChallengeServiceImpl implements ChallengeService {
         }
 
         if (currentIndex == -1) {//FINAL
-            challengeDetailDto.setCurrentPhase(ChallengePhaseEnum.FINAL);
-            challengeDetailDto.setNextPhase(ChallengePhaseEnum.FINAL);
+            challengeDetailDto.setCurrentPhase(FINAL);
+            challengeDetailDto.setNextPhase(FINAL);
         } else {
             challengeDetailDto.setCurrentPhase(CHALLENGE_TIMELINE[currentIndex]);
             challengeDetailDto.setNextPhase(CHALLENGE_TIMELINE[nextIndex > -1 ? nextIndex : currentIndex]);
         }
     }
 
-    public ChallengeRegistrantDto acceptRegistrant(Long registrantId, ChallengePhaseEnum activePhase) {
-        ChallengeRegistrantEntity registrant = challengeRegistrantRepository.findOne(registrantId);
-
-        if (registrant != null && activePhase != registrant.getActivePhase()) {
-            registrant.setActivePhase(activePhase);
-            registrant.setDisqualified(null);
-            registrant.setDisqualifiedReason(null);
-            registrant = challengeRegistrantRepository.save(registrant);
-        }
-
-        return dozerMapper.map(registrant, ChallengeRegistrantDto.class);
-    }
-
-    @Override
-    public List<ChallengeRegistrantFunnelItem> getChallengeRegistrantFunnel(Long challengeId, String ownerEmail) {
-        List<ChallengeRegistrantFunnelItem> funnel = new ArrayList<>();
-        ChallengeDto challengeDto = findChallengeById(challengeId, ownerEmail);
-        Map<ChallengePhaseEnum, ChallengeRegistrantPhaseItem> numberOfRegistrantsByPhase =
-                challengeRegistrantService.countNumberOfRegistrantsByPhase(challengeId);
-        Map<ChallengePhaseEnum, ChallengeSubmissionPhaseItem> numberOfSubmissionsByPhase =
-                challengeSubmissionService.countNumberOfSubmissionsByPhase(challengeId, null);
-        Map<ChallengePhaseEnum, ChallengeSubmissionPhaseItem> numberOfUnreadSubmissionsByPhase =
-                challengeSubmissionService.countNumberOfSubmissionsByPhase(challengeId, Boolean.FALSE);
-
-        for (Map.Entry<ChallengePhaseEnum, ChallengeRegistrantPhaseItem> entry : numberOfRegistrantsByPhase.entrySet()) {
-            ChallengePhaseEnum phase = entry.getKey();
-            Long participant = entry.getValue().getRegistration();
-            Long submission = 0L;
-            Long unreadSubmission = 0L;
-            if (numberOfSubmissionsByPhase.get(phase) != null) {
-                submission = numberOfSubmissionsByPhase.get(phase).getSubmission();
-            }
-            if (numberOfUnreadSubmissionsByPhase.get(phase) != null) {
-                unreadSubmission = numberOfUnreadSubmissionsByPhase.get(phase).getSubmission();
-            }
-
-            if (isValidPhase(challengeDto, phase)) {
-                funnel.add(new ChallengeRegistrantFunnelItem(phase, participant, submission, unreadSubmission));
-            }
-        }
-
-        Long numberOfFinalists = challengeRegistrantService.countNumberOfFinalists(challengeId);
-        Long numberOfWinners = Long.valueOf(challengeRegistrantService.countNumberOfWinners(challengeId));
-        funnel.add(new ChallengeRegistrantFunnelItem(ChallengePhaseEnum.WINNER, numberOfFinalists, numberOfWinners, 0L));
-
-        Comparator<ChallengeRegistrantFunnelItem> sortByPhaseComparator = (item1, item2) ->
-                item1.getPhase().getOrder() - item2.getPhase().getOrder();
-        return funnel.stream().sorted(sortByPhaseComparator).collect(toList());
-    }
-
-    private boolean isValidPhase(ChallengeDto challengeDto, ChallengePhaseEnum phase) {
-        switch (phase) {
-            case REGISTRATION:
-                return StringUtils.isNotEmpty(challengeDto.getRegistrationDate());
-            case IDEA:
-                return StringUtils.isNotEmpty(challengeDto.getIdeaSubmissionDate());
-            case UIUX:
-                return StringUtils.isNotEmpty(challengeDto.getUxSubmissionDate());
-            case PROTOTYPE:
-                return StringUtils.isNotEmpty(challengeDto.getPrototypeSubmissionDate());
-            case FINAL:
-                return StringUtils.isNotEmpty(challengeDto.getSubmissionDate());
-            default:
-                return false;
-        }
-    }
-
-    @Override
-    public List<ChallengeRegistrantDto> qualifyAllRegistrants(String ownerEmail, ChallengeQualificationDto challengeQualificationDto) {
-        List<ChallengeRegistrantDto> qualifiedRegistrants = new ArrayList<>();
-        ChallengePhaseEnum qualifyingPhase = challengeQualificationDto.getNextPhase();
-
-        for (Long registrantId : challengeQualificationDto.getRegistrantIds()) {
-            ChallengeRegistrantDto registrantDto = acceptRegistrant(registrantId, qualifyingPhase);
-            if (registrantDto.getActivePhase() == qualifyingPhase) {
-                qualifiedRegistrants.add(registrantDto);
-            }
-        }
-        return qualifiedRegistrants;
-    }
-
-    public ChallengeRegistrantDto rejectRegistrant(String ownerEmail, RejectRegistrantDto rejectRegistrantDto){
-        Iterator<ChallengeRegistrantEntity> registrantIter = challengeRegistrantRepository.search(QueryBuilders.termQuery("registrantId", rejectRegistrantDto.getRegistrantId())).iterator();
-        if (!registrantIter.hasNext()) return null;
-
-        ChallengeRegistrantEntity registrant = registrantIter.next();
-        ChallengeEntity challenge = challengeRepository.findOne(registrant.getChallengeId());
-        if (!ownerEmail.equalsIgnoreCase(challenge.getAuthorEmail())) {
-            return null;
-        }
-
-        registrant.setDisqualified(Boolean.TRUE);
-        registrant.setDisqualifiedReason(rejectRegistrantDto.getReason());
-        registrant = challengeRegistrantRepository.save(registrant);
-        return dozerMapper.map(registrant, ChallengeRegistrantDto.class);
-    }
 }
